backend/requirements.txt
# Atlas Network Monitor Backend Dependencies

# FastAPI and related
python-multipart==0.0.6

psutil==5.9.6
# Additional utilities
python-dotenv==1.0.0
annotated-types==0.7.0
anyio==4.10.0
async-timeout==5.0.1
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
pyp0f==0.3.0
redis==6.4.0
scapy==2.6.1
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
oauthlib==3.2.2
requests==2.31.0
